Upload current versions of server files
I've placed some comments in the python file with an idea of how i want to extract the variables from the JSON, but i'm having trouble getting it to work inside the framework of the python http handler libraries that the code uses. Any help would be more than welcome.